fixed different type comparison

+++ src/util/Kokkos_grow.hpp
@@ -12,7 +12,7 @@ void grow(Kokkos::View<T, P...>& view, idx_t size, real_t safetyMargin = 1.1_r)
 {
     assert(size >= 0);
 
-    if (view.extent(0) <= size)
+    if (view.extent(0) <= static_cast<unsigned long>(size))
     {
         Kokkos::resize(view, idx_c(real_c(size) * safetyMargin));
     }
